backend/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
eventlet==0.37.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
pymongo>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
